feat: refactor dtos with mapped types
