--- src/users/domain/dto/create-role.dto.ts
@@ -1,14 +1,0 @@
-import {IsNotEmpty, IsString} from "class-validator";
-import {ApiProperty} from "@nestjs/swagger";
-
-export class CreateRoleDto {
-    @ApiProperty()
-    @IsNotEmpty()
-    @IsString()
-    name: string;
-
-    @ApiProperty()
-    @IsNotEmpty()
-    @IsString()
-    description: string;
-}
