--- src/users/domain/dto/update-role.dto.ts
@@ -1,16 +1,0 @@
-/*import {IsOptional, IsString} from 'class-validator';
-
-export class UpdateRoleDto {
-    @IsString()
-    @IsOptional()
-    name?: string;
-
-    @IsString()
-    @IsOptional()
-    description?: string;
-}
-*/
-import { PartialType } from "@nestjs/swagger";
-import { CreateRoleDto } from "./create-role.dto";
-
-export class UpdateRoleDto extends PartialType(CreateRoleDto) {}
