--- src/users/domain/dto/update-user-profile.dto.ts
@@ -1,19 +1,0 @@
-import {IsNotEmpty, IsString} from "class-validator";
-
-export class UpdateUserProfileDto {
-    @IsNotEmpty()
-    @IsString()
-    dni: string;
-
-    @IsNotEmpty()
-    @IsString()
-    first_name: string;
-
-    @IsNotEmpty()
-    @IsString()
-    last_name: string;
-
-    @IsNotEmpty()
-    @IsString()
-    phone_number: string;
-}
