--- src/users/domain/dto/update-user.dto.ts
@@ -1,2 +1,0 @@
-export class UpdateUserDto {
-}
