+++ src/users/domain/dto/user.dto.ts
@@ -1,0 +1,16 @@
+import { ApiProperty, PartialType } from "@nestjs/swagger";
+import { IsNotEmpty, IsString } from "class-validator";
+
+export class CreateUserDto {
+    @ApiProperty()
+    @IsNotEmpty()
+    @IsString()
+    email: string;
+
+    @ApiProperty()
+    @IsNotEmpty()
+    @IsString()
+    password_hash: string;
+}
+
+export class UpdateUserDto extends PartialType(CreateUserDto) {}
